changes to whitespace handling of formatted code
I have the weird idea to ask stog to handle such code formatted
*inside* markdown documents, which are later processed by markdown. It
only works well if the HTML produced by stog is recognized as raw HTML
by markdown, and for this block content needs to be separated from the
rest of the document by a blank line (before and after). I explicitely
added this blank line ("\n\n") on both side of the formatted
result. This does not change the semantics of the HTML document, so
I think the change is safe for other users -- one may even argue that
it formats the produced HTML in a way more comfortable to the human
eye.