support for various publication levels instead of a published / not published dichotomy
My blog workflow uses two building rule: a "make site" target that
builds the final blog published to the readers, and a "make test"
target that builds a test instance of the blog at a private URL, used
to check rendering before publication and ask review from coauthors.

If an element is put in "draft" mode, it will be visible from the test
instance (rendered with --publication-level=draft), but not from the
release instance. I can have some posts in draft mode, and at the same
time make changes to released posts and push them to the release
server, the drafts won't be included.

There is a bug in the interaction between publication level and
caching: elements are cached irrespectively of their publication
status, so putting an element back to "published=false" or
"published=draft" when it was cached with "published=true" will not
remove it from the cached rendering. Note that this bug is due to the
cache logic not taking all the inputs (here configuration variables)
into account, and was not introduced by this patch.